Orders, products and the stuffz

+++ src/main/kotlin/xyz/nygaard/store/user/Token.kt
@@ -4,7 +4,7 @@ import com.github.nitram509.jmacaroons.Macaroon
 import xyz.nygaard.extractUserId
 import java.util.*
 
-class Token(val macaroon: Macaroon, val balance: Int, val revoked: Boolean) {
+class Token(val macaroon: Macaroon, val balance: Long, val revoked: Boolean) {
     val id = macaroon.extractUserId()
 
     fun toDTO() = TokenResponse(id, balance)
